Partial fix for Solution Saving Order + 1 more
Partially fixed issue #28 - solutions are no longer jumbled, but are still not appearing in order. Each one of cube solutions is correct though and they can be searched through with CMD-F so this works fine for now but needs fixing fully before v0.11 is completed. It should be noted that enabling solution saving in the developer settings menu now causes a small performance decrease, so best to leave it OFF when not required (OFF is the default state for this setting anyway).

Removed a duplicate line of code from solve.py in the totalSolve function which was causing the fixAlignedCenters step to be run twice with each solve, thereby resulting in a non-trivial performance improvement.